Rename Server --> Publisher
