+++ app/Publisher/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required (VERSION 3.0)
 
-project(Server)
+project(Publisher)
 
 # Dependencies
 set(DEPENDENCIES leptonAPI)
