+++ app/Publisher/Common.h
